Rename the IsActive test resources
The idea of a "complex" is no longer applicable so we need to rename all
the relevant test resources.
