+++ packages/harpstrata/src/harpstrata/HarpStrata/testResources/index.ts
@@ -1,6 +1,6 @@
 import { PozitionIds } from '../../Pozition'
 import { PitchIds, EXAMPLE_PITCH_MATRICES } from '../../Pitch'
-import { EXAMPLE_IS_ACTIVE_COMPLEXES } from '../../IsActive'
+import { EXAMPLE_IS_ACTIVE_IDS_PAIR } from '../../IsActive'
 import type { HarpStrata } from '../../HarpStrata'
 import { EXAMPLE_DEGREE_MATRICES } from '../../Degree'
 import { MAJOR_DIATONIC_APPARATUS } from '../../Apparatus'
@@ -10,10 +10,10 @@ const C_MAJOR_DIATONIC_FIRST_POZITION_C_MAJOR_PENTATONIC: HarpStrata = {
   degreeMatrix: EXAMPLE_DEGREE_MATRICES.MAJOR_DIATONIC_FIRST_POZITION,
   pitchMatrix: EXAMPLE_PITCH_MATRICES.MAJOR_DIATONIC_C_HARMONICA,
   activeDegreeIds:
-    EXAMPLE_IS_ACTIVE_COMPLEXES
+    EXAMPLE_IS_ACTIVE_IDS_PAIR
       .C_MAJOR_DIATONIC_FIRST_POZITION_C_MAJOR_PENTATONIC.activeDegreeIds,
   activePitchIds:
-    EXAMPLE_IS_ACTIVE_COMPLEXES
+    EXAMPLE_IS_ACTIVE_IDS_PAIR
       .C_MAJOR_DIATONIC_FIRST_POZITION_C_MAJOR_PENTATONIC.activePitchIds,
   pozitionId: PozitionIds.First,
   rootPitchId: PitchIds.C,
@@ -25,10 +25,10 @@ const C_MAJOR_DIATONIC_SECOND_POZITION_G_MAJOR_PENTATONIC: HarpStrata = {
   degreeMatrix: EXAMPLE_DEGREE_MATRICES.MAJOR_DIATONIC_SECOND_POZITION,
   pitchMatrix: EXAMPLE_PITCH_MATRICES.MAJOR_DIATONIC_C_HARMONICA,
   activeDegreeIds:
-    EXAMPLE_IS_ACTIVE_COMPLEXES
+    EXAMPLE_IS_ACTIVE_IDS_PAIR
       .C_MAJOR_DIATONIC_SECOND_POZITION_G_MAJOR_PENTATONIC.activeDegreeIds,
   activePitchIds:
-    EXAMPLE_IS_ACTIVE_COMPLEXES
+    EXAMPLE_IS_ACTIVE_IDS_PAIR
       .C_MAJOR_DIATONIC_SECOND_POZITION_G_MAJOR_PENTATONIC.activePitchIds,
   pozitionId: PozitionIds.Second,
   rootPitchId: PitchIds.G,
