+++ packages/harpstrata/src/harpstrata/IsActive/ActiveIds/testResources/index.ts
@@ -24,8 +24,17 @@ const G_MAJOR_PENTATONIC_PITCH_IDS: ActivePitchIds = [
   PitchIds.G,
 ] as const
 
-export const EXAMPLE_ACTIVE_IDS = {
-  MAJOR_PENTATONIC_SCALE_DEGREE_IDS,
-  C_MAJOR_PENTATONIC_PITCH_IDS,
-  G_MAJOR_PENTATONIC_PITCH_IDS,
+const C_MAJOR_DIATONIC_FIRST_POZITION_C_MAJOR_PENTATONIC = {
+  activeDegreeIds: MAJOR_PENTATONIC_SCALE_DEGREE_IDS,
+  activePitchIds: C_MAJOR_PENTATONIC_PITCH_IDS,
+} as const
+
+const C_MAJOR_DIATONIC_SECOND_POZITION_G_MAJOR_PENTATONIC = {
+  activeDegreeIds: MAJOR_PENTATONIC_SCALE_DEGREE_IDS,
+  activePitchIds: G_MAJOR_PENTATONIC_PITCH_IDS,
+} as const
+
+export const EXAMPLE_IS_ACTIVE_IDS_PAIR = {
+  C_MAJOR_DIATONIC_FIRST_POZITION_C_MAJOR_PENTATONIC,
+  C_MAJOR_DIATONIC_SECOND_POZITION_G_MAJOR_PENTATONIC,
 } as const
