+++ packages/harpstrata/src/harpstrata/IsActive/index.ts
@@ -4,4 +4,4 @@ export type { ActiveDegreeIds, ActivePitchIds, ActiveIds } from './ActiveIds'
 
 export { IsActiveIds } from './types'
 
-export { EXAMPLE_IS_ACTIVE_COMPLEXES } from './testResources'
+export { EXAMPLE_IS_ACTIVE_IDS_PAIR } from './ActiveIds'
