--- packages/harpstrata/src/harpstrata/IsActive/testResources/index.ts
@@ -1,22 +1,0 @@
-import { EXAMPLE_ACTIVE_IDS } from '../ActiveIds'
-
-const {
-  MAJOR_PENTATONIC_SCALE_DEGREE_IDS,
-  C_MAJOR_PENTATONIC_PITCH_IDS,
-  G_MAJOR_PENTATONIC_PITCH_IDS,
-} = EXAMPLE_ACTIVE_IDS
-
-const C_MAJOR_DIATONIC_FIRST_POZITION_C_MAJOR_PENTATONIC = {
-  activeDegreeIds: MAJOR_PENTATONIC_SCALE_DEGREE_IDS,
-  activePitchIds: C_MAJOR_PENTATONIC_PITCH_IDS,
-} as const
-
-const C_MAJOR_DIATONIC_SECOND_POZITION_G_MAJOR_PENTATONIC = {
-  activeDegreeIds: MAJOR_PENTATONIC_SCALE_DEGREE_IDS,
-  activePitchIds: G_MAJOR_PENTATONIC_PITCH_IDS,
-} as const
-
-export const EXAMPLE_IS_ACTIVE_COMPLEXES = {
-  C_MAJOR_DIATONIC_FIRST_POZITION_C_MAJOR_PENTATONIC,
-  C_MAJOR_DIATONIC_SECOND_POZITION_G_MAJOR_PENTATONIC,
-} as const
